Allow to track time for the lifespan of a process
When started with extra arguments, wtrd(1) will spawn this command and
record time spent on projects for the lifetime of this process.

This allows to rely on wtrd(1) when starting a user session e.g. in
.xinitrc:

```
exec /usr/local/bin/wtrd /usr/local/bin/dbus-launch --exit-with-session /usr/local/bin/fvwm3
```