Only show relevent month/year in graph
When generating graph, the last column displayed the month/year if they
changed beween the start and the end of the week.  Depending on the
requested graph, the last shown date might be earlier that the last day
of the week, and the last column may only show days in the current
month/year.  In this case, there is no point in displaying the next
month/year names which has no data in the grah.