Improve database merging
Instead of removing all known entries for imported hosts and re-creating
them, attempt to add the new values, and on conflict, only update the
old value if the recorded time is less than what is presented in the
import database.

This will prevent loosing information on host D when importing database
from host B and C which both import from host A: B and C might have a
different view of the time spent on host A (because they where not
imported simultaneously), and the order of imports on host D should have
predictible results.