Show per-project activity
Active projects where marked with a single '+', giving no information
about the number of process running and accounting for them.

Track the number of processes active for each project, and show this
number as a bar to help distinguish projects with a low number of
processes from those with a large number of processes.