Add a legend in graph output
The color indication help figure out the amount of time per day
relatively to other days, but lacks indications about the amount of time
each color correspond to.

Improve this by printing the minimum and maximum values with the
corresponding colors.

In the future, we might implement some kind of gradient with
intermediate values.